# src/Nullref.FullStackDemo.Frontend/src/app/api/models/id-response-model.ts
/* tslint:disable */
/* eslint-disable */
export interface IdResponseModel {
  id: string;
}

export const IdResponseModelMetadata = {
  identifier: {
    id: 'id',
  },
  maxLength: {
  },
  minLength: {
  },
  readOnly: {
    id: false,
  },
  required: {
    id: true,
  },
  nullable: {
    id: false,
  },
  default: {
  },
  format: {
    id: 'uuid',
  },
  allowSort: {
  },
  displayName: {
  },
  description: {
  },
}